+++ src/test/java/updateExistingPet.java
@@ -6,7 +6,7 @@ import org.junit.Test;
 
 public class updateExistingPet {
     PetEndpoint petEndpoint=new PetEndpoint();
-    int  createdPetId;
+    long  createdPetId;
 
     @Before
     public void createPreconditions() {
